CA-3376: Updated launch.json to allow walking through TS spec files.
While using vscode and the current file open in the editor is a spec file with mocha tests, select the "Current TS Spec File" launch config from the debug menu and click run (or hit F5). This launch config allows the use of break points and the ability to walk through TypeScript files directly.