CA-3423: OrdersService now automatically gets all pages of orders by using the returned pagination from the CA API
When the pagination property is present on the response body when making a call to the /orders endpoint the OrderService checks to see if lastPage is true. If it is not then another call is made to the /orders endpoint with the same order request that the user sent, however the lastKey property is now set. Once lastPage is true the concatenated orders array from all calls from pages are returned.