Updated ActionsService spec to ensure that the actually returned javascript object converts the correct strings into Dates()
Instead of yielding actual Action types when specing out the ActionsService the specs now yield plain JavaScript objects, as those are what are actually returned by the request client. This also allows us to test the endDate and ensure if it is present in the response from the client it is converted to a Date().