CA-4279 Removing e2e test for Business/es retrieval
We opted to remove these e2e tests because the current Travis build is using an API key instead of a sessionId. Currently there is no way to retrieve businesses with an API key. We could add a sessionId env var but this would require extra time since the env var would need to be kept fresh with an updated cron job. Another solution could be to simply create a user that only belongs to the Production Test business and set the login and password as env vars.